Extracting BlogItem from home and added routing to details page
